CNN-Transfer-Learning/tutorial/requirements.txt
tensorflow
tensorflow-gpu
pandas
Pillow
keras
flask